implement quick sort
